barchart 통계자료 탄소 배출량으로 변경

+++ app/src/main/java/com/onehundredyo/batteryfreeze/fragment/DataUsage.kt
@@ -2,5 +2,5 @@ package com.example.myapplication
 
 data class DataUsage(
     val title:String,
-    val datausage: Long,
+    val datausage: Double,
 )
